fix(syntax): Don’t falsely set `inFunction=false` inside LSC expressions
Originally, this was designed to prevent the use of `return` inside expression bodies that might get converted to IIFEs. However, it had the unintended side effect of banning `await` as well as fooling other parts of the parser.

Instead, we introduce a new `state.banInvalidIifeStatements` flag that bans only the statements that break IIFEs.